refactor(timer): extract timer widget
